// src/map-interactable-controller/enums.ts
export enum MAP_INTERACTABLE_CONTROLLER_ENUMS {
  MAP_1_NPC_1_OFFSET_X = 655,
  MAP_1_NPC_1_OFFSET_Y = 435,
  MAP_1_NPC_2_OFFSET_X = 655,
  MAP_1_NPC_2_OFFSET_Y = 635,
  MAP_1_COMPUTER_1_OFFSET_X = 845,
  MAP_1_COMPUTER_1_OFFSET_Y = 150,
  MAP_1_COMPUTER_1_INTERACTION_TILE_ID = 1817,
  MAP_1_COMPUTER_2_OFFSET_X = 1035,
  MAP_1_COMPUTER_2_OFFSET_Y = 150,
  MAP_1_COMPUTER_2_INTERACTION_TILE_ID = 1818,
}